# specs/001-task-crud-auth/contracts/tasks.openapi.yaml
openapi: 3.0.3
info:
  title: Tasks API
  description: API for managing user tasks with full CRUD operations
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

paths:
  /api/tasks:
    get:
      summary: Get all user tasks
      description: Retrieves all tasks belonging to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      summary: Create a new task
      description: Creates a new task for the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Task title
                  example: "Buy groceries"
                description:
                  type: string
                  maxLength: 1000
                  description: Optional task description
                  example: "Milk, eggs, bread, fruits"
      responses:
        '201':
          description: Task successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/tasks/{taskId}:
    get:
      summary: Get a specific task
      description: Retrieves a specific task belonging to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to retrieve
      responses:
        '200':
          description: Successfully retrieved task
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update a task
      description: Updates all fields of a specific task belonging to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Task title
                  example: "Updated task title"
                description:
                  type: string
                  maxLength: 1000
                  description: Optional task description
                  example: "Updated task description"
                completed:
                  type: boolean
                  description: Completion status
                  example: false
      responses:
        '200':
          description: Task successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      summary: Partially update a task
      description: Updates specific fields of a task belonging to the authenticated user (useful for toggling completion status)
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Task title
                  example: "Updated task title"
                description:
                  type: string
                  maxLength: 1000
                  description: Optional task description
                  example: "Updated task description"
                completed:
                  type: boolean
                  description: Completion status
                  example: true
      responses:
        '200':
          description: Task successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete a task
      description: Deletes a specific task belonging to the authenticated user
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the task to delete
      responses:
        '204':
          description: Task successfully deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from authentication endpoint

  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns this task
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          description: Task title
          example: "Buy groceries"
        description:
          type: string
          description: Optional task description
          example: "Milk, eggs, bread, fruits"
        completed:
          type: boolean
          description: Completion status
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2023-01-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Task last update timestamp
          example: "2023-01-01T12:00:00Z"
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Title is required and must be between 1-200 characters"
    UnauthorizedError:
      description: Unauthorized - invalid or missing JWT token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Authentication required"
    NotFoundError:
      description: Resource not found - could be the task doesn't exist or belongs to another user
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Task not found"

security:
  - bearerAuth: []